requirements.txt

spotipy==2.23.0
pandas==2.0.3
python-dotenv==0.21.0
